extract attribute settings module from the model settings module
